+++ egui/src/context.rs
@@ -29,7 +29,7 @@ struct Options {
 pub(crate) struct FrameState {
     /// Starts off as the screen_rect, shrinks as panels are added.
     /// The `CentralPanel` does not change this.
-    /// This is the area avilable to Window's.
+    /// This is the area available to Window's.
     available_rect: Rect,
 
     /// Starts off as the screen_rect, shrinks as panels are added.
@@ -93,7 +93,7 @@ impl FrameState {
 
     pub(crate) fn allocate_central_panel(&mut self, panel_rect: Rect) {
         // Note: we do not shrink `available_rect`, because
-        // we alllow windows to cover the CentralPanel.
+        // we allow windows to cover the CentralPanel.
         self.unused_rect = Rect::nothing(); // Nothing left unused after this
         self.used_by_panels = self.used_by_panels.union(panel_rect);
     }
@@ -196,15 +196,27 @@ impl CtxRef {
     /// Use `ui.interact` instead
     pub(crate) fn interact(
         &self,
-        layer_id: LayerId,
         clip_rect: Rect,
         item_spacing: Vec2,
-        rect: Rect,
+        layer_id: LayerId,
         id: Id,
+        rect: Rect,
         sense: Sense,
     ) -> Response {
         let interact_rect = rect.expand2((0.5 * item_spacing).min(Vec2::splat(5.0))); // make it easier to click
-        let hovered = self.contains_mouse(layer_id, clip_rect, interact_rect);
+        let hovered = self.rect_contains_mouse(layer_id, clip_rect.intersect(interact_rect));
+        self.interact_with_hovered(layer_id, id, rect, sense, hovered)
+    }
+
+    /// You specify if a thing is hovered, and the function gives a `Response`.
+    pub(crate) fn interact_with_hovered(
+        &self,
+        layer_id: LayerId,
+        id: Id,
+        rect: Rect,
+        sense: Sense,
+        hovered: bool,
+    ) -> Response {
         let has_kb_focus = self.memory().has_kb_focus(id);
 
         // If the the focus is lost after the call to interact,
@@ -215,6 +227,7 @@ impl CtxRef {
             // Not interested or allowed input:
             return Response {
                 ctx: self.clone(),
+                layer_id,
                 id,
                 rect,
                 sense,
@@ -241,6 +254,7 @@ impl CtxRef {
             if hovered {
                 let mut response = Response {
                     ctx: self.clone(),
+                    layer_id,
                     id,
                     rect,
                     sense,
@@ -273,6 +287,7 @@ impl CtxRef {
                 // miss
                 Response {
                     ctx: self.clone(),
+                    layer_id,
                     id,
                     rect,
                     sense,
@@ -288,6 +303,7 @@ impl CtxRef {
             let clicked = hovered && active && self.input.mouse.could_be_click;
             Response {
                 ctx: self.clone(),
+                layer_id,
                 id,
                 rect,
                 sense,
@@ -301,6 +317,7 @@ impl CtxRef {
         } else if self.input.mouse.down {
             Response {
                 ctx: self.clone(),
+                layer_id,
                 id,
                 rect,
                 sense,
@@ -314,6 +331,7 @@ impl CtxRef {
         } else {
             Response {
                 ctx: self.clone(),
+                layer_id,
                 id,
                 rect,
                 sense,
@@ -652,8 +670,7 @@ impl Context {
         self.memory().layer_id_at(pos, resize_grab_radius_side)
     }
 
-    pub fn contains_mouse(&self, layer_id: LayerId, clip_rect: Rect, rect: Rect) -> bool {
-        let rect = rect.intersect(clip_rect);
+    pub(crate) fn rect_contains_mouse(&self, layer_id: LayerId, rect: Rect) -> bool {
         if let Some(mouse_pos) = self.input.mouse.pos {
             rect.contains(mouse_pos) && self.layer_id_at(mouse_pos) == Some(layer_id)
         } else {